Collapse test_inner upward
Resolves https://github.com/tokio-rs/tokio/pull/2107#discussion_r371987618
